- removed compile warning (signed vs. unsigned, float vs. double & bool vs. BOOL) - removed deprecation warning at usage of 'fopen' (now 'fopen_s' is used) - removed deprecation warning at usage of 'sscanf' (now 'sscanf_s' is used)
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@105 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/AABBUtil.cpp
@@ -23,8 +23,8 @@
 
 D3DXVECTOR3 WorldPickRay(float x, float y, const LPD3DXMATRIX mProj, const LPD3DXMATRIX mView)
 {
-	x = (x*2.0-1.0)/mProj->_11;
-	y = (y*2.0-1.0)/mProj->_22;
+	x = float((x*2.0-1.0)/mProj->_11);
+	y = float((y*2.0-1.0)/mProj->_22);
 	D3DXVECTOR3 pick(x, 1.0f, y);
 	D3DXMATRIX mViewI;
 	D3DXMatrixInverse(&mViewI, NULL, (const D3DXMATRIX *)&mView);
